Include logging for Github Actions and change depth

+++ src/data_cleanup.py
@@ -2,8 +2,16 @@
 # -*- coding: utf-8 -*-
 
 import shutil
+import os
+from base_logger import logger
+
+logger.info("[DATA-CLEANUP]- Start data folder deletion")
 
 if __name__ == "__main__":
-    shutil.rmtree('data/')
+    if os.path.exists("data/"):
+        logger.info("[DATA-CLEANUP]- Data folder exists")
+        shutil.rmtree("data/")
+    else:
+        logger.info("[DATA-CLEANUP]- Data folder does not exists. Skip.")
 
-
+logger.info("[DATA-CLEANUP]- Data folder deletion finished")
